creating a side table that save Scenario Inputs
The SCENARIOS table matches the DS_FINAL table on column ScenarioName.  The SCENARIOS tables has the name/value pair for all macro variables definited in the macro.

This will be a helpful step forward as we add many more scenarios.  It will also allow evaluation to prevent re-running the same scenario.