Major Update
SIR.sas is now COVID_19.sas . This reflect the inclusion of multipel ModelType=.  There are two SIR models (PROC TMODEL and DS) and a SEIR model (PROC TMODEL).

There are 12 scenario calls to the macro.  All store results in MODEL_FINAL.  The Scenario parameters are stores in SCENARIOS.  These are linked by ScenarioIndex.

Option parameter on the %EASYRUN macro of plots=YES will turn on SGPLOT for review of results and quick comparison of model types