plots=yes works for previously run scenarios
The logic is updated to display plots for scenarios that are identified as previously run.  This is done withour rerunning the scenario.  Instead, the ScenarioIndex for the most recentl match scenario is use to recall the data from STORE.MODEL_FINAL for the plotting.  This logic can also extend to the application level this code will be embedded in.