FIX: create ISOChangeEvent and EventY* columns for each scenario
Fixed the issue where a first scenario may not have values for ISOChangeDate, ISOChangeEvent and SocialDistancingChange.  In these cases the columns for ISOChangeEvent and EventY_* would not be created in MODEL_FINAL causing subsequent scenarios to not add this information to the MODEL_FINAL table.  Now these columns are created and made empty for scenarios without the change inputs.