Fitting Method Finalized
The fitting method is updated to not use manual inputs for SocialDistancingChange.  It uses FIT_INPUT data to estiamte R0 at the start of time, the time at which R0 changes, called CURVEBEND1 and the new R0 after CURVEBEND1  is call R0_BEND_FIT.

This fix is happening on this branch due to its consideration and research happening here.  While it is not impacted by ISOChangeWindow values it will impact all COVID_19.sas files that include the fitting method.

Future develpment will include updating this to estimate more than 1 change point for R0.